Rename store to state
