+++ src/features/timeEntry/state/issues/types.ts
@@ -1,6 +1,6 @@
 import { Issue } from 'shared/types';
-import { AppState } from 'store';
-import { ThunkAction } from 'store/thunk';
+import { AppState } from 'state';
+import { ThunkAction } from 'state/thunk';
 
 export type IssueFilterValue = {
     status?: string,
