+++ src/features/timeEntry/state/timing/types.ts
@@ -1,6 +1,6 @@
 import { Dispatch, Action } from 'redux';
-import { AppState } from 'store';
-import { ThunkAction } from 'store/thunk';
+import { AppState } from 'state';
+import { ThunkAction } from 'state/thunk';
 import { TimesheetEntry } from 'shared/types';
 
 export type TimingsFormState = {
