+++ src/index.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 import { render } from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import { store } from './store/createStore';
+import { store } from './state/createStore';
 
 import { AppRoot } from './containers/appRoot';
